+++ www/src/pages/docs/actions.js
@@ -8,7 +8,6 @@ import { sortBy } from "lodash-es"
 import APIReference from "../../components/api-reference"
 import Layout from "../../components/layout"
 import Container from "../../components/container"
-import { itemListDocs } from "../../utils/sidebar/item-list"
 
 class ActionCreatorsDocs extends React.Component {
   render() {
@@ -25,7 +24,7 @@ class ActionCreatorsDocs extends React.Component {
     )
 
     return (
-      <Layout location={this.props.location} itemList={itemListDocs}>
+      <Layout location={this.props.location}>
         <Container>
           <Helmet>
             <title>Actions</title>
@@ -41,8 +40,13 @@ class ActionCreatorsDocs extends React.Component {
             <a href="http://redux.js.org">Redux</a>
             {` `}
             internally to manage state. When you implement a Gatsby API, you are
-            passed a collection of actions (equivalent to boundActionCreators in
-            redux) which you can use to manipulate state on your site.
+            passed a collection of actions (equivalent to actions bound with
+            {` `}
+            <a href="https://redux.js.org/api/bindactioncreators">
+              bindActionCreators
+            </a>
+            {` `}
+            in redux) which you can use to manipulate state on your site.
           </p>
           <p>
             The object