+++ packages/gatsby-source-drupal/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.17...gatsby-source-drupal@3.3.18) (2020-02-01)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.16...gatsby-source-drupal@3.3.17) (2020-01-29)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.15...gatsby-source-drupal@3.3.16) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.13...gatsby-source-drupal@3.3.15) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.13...gatsby-source-drupal@3.3.14) (2020-01-09)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.12...gatsby-source-drupal@3.3.13) (2019-12-20)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.10...gatsby-source-drupal@3.3.12) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
+## [3.3.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.10...gatsby-source-drupal@3.3.11) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-source-drupal
+
 ## [3.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-drupal@3.3.9...gatsby-source-drupal@3.3.10) (2019-12-05)
 
 ### Bug Fixes